+++ flake.nix
@@ -55,6 +55,7 @@
               };
               "ipykernel".ipykernel = { };
               "ihaskell".ihaskell = { };
+              "julia".ijulia = { };
             };
           };
         };